Select data using the scatterplot
I kept getting this error when using the selectedData attribute
from the scatterplot to update the graphs and table:
    unrecognized GUI edit: selections[0].xref
The error only occurs if the scatterplot data is updated as an
output after selecting points from the scatterplot, not if the
points are updated after selecting via checkboxes or pie graphs.

There is only one bug report for this error:
https://github.com/plotly/react-plotly.js/issues/290
It suggests that using versions of plotly.js prior to 2.13.2 solve
the problem.

I tried plotly v5.11.0, 5.13.1, 5.9.0, and 5.8.2, which use
earlier versions of plotly.js.  None of them solved the error.

In the Dash documentation for interactive graphing, they
demonstrate using selectedData with a crossfilter example:
https://dash.plotly.com/interactive-graphing#generic-crossfilter-recipe
This example worked for me.

By gradually simplifying the code in the crossfiltering example,
it seems that you need to update the selection mode after creating
a new figure using selectedData:
fig.update_layout(
    newselection_mode="gradual",
)
Then using selectedData works.

The example code from the Dash documentation includes a link to
these release notes for Dash, describing other cool features of
using selectedData:
https://medium.com/@plotlygraphs/notes-from-the-latest-plotly-js-release-b035a5b43e21

I did not implement the built-in option to colour points based on
whether or not they were selected, but it might be cool for future.