Bug fix: clicking on a plot section sequentially doesn't fire callback
Clicking the same figure section in a pie chart or bar graph multiple
times in a row doesn't fire the callback.  So if a section was clicked,
then the reset button was clicked, then the same section was clicked
again, nothing would happen.

The workaround is to reset the figure's clickData property to None
after each callback.

This may cause problems in future if we need to increment clickData,
e.g. to keep track of how many times something has been selected.

The bug and workaround are described here:
https://github.com/plotly/dash/issues/1300