Display figures in carousel
Note that this defaults to appearing like a dbc.CardGroup when the
total number of figures to display is the same as the maximum number
of figures the carousel displays at one time.

The benefit of a carousel over a CardGroup is that I can't figure
out how to dynamically change the number of rows in a CardGroup to
adapt to more figures.

Also if there are many figures, it may not be desireable to display
them all at once, because you would then have to scroll down quite
far to get to the data table.

Dash Bootstrap Components also has a carousel component, however I
can't figure out how to display multiple items at once with it.

The dash_slick package doesn't natively support displaying
thumbnails.  One workaround suggested online is to use two linked
carousels -- one to display fewer larger images, and the other to
display more small images.