Change SVG of chip icon [NOT WORKING; WILL ABANDON]
SVG is too large and is displayed outside of the chip (check this
with the inspector function of a browser).

I can't figure out how to rescale and translate the SVG, so will
abondon using the Mantine chips.  This commit is in case we need
to know later what I tried.