Make dataframe filtering robust to list values in dataframe
Filter criteria are provided as a list of values.  The previous
filtering method used the isin() method of a dataframe to find
cells matching an item in the filter criteria list. However this
method doesn't work with list inputs for the cells, so if a cell
contained a list the filtering always returned False even if the
list contained a value in the filter criteria.

This StackOverflow answer was helpful to find a solution:
https://stackoverflow.com/a/53343080