Use Dash Bootstrap table instead of Dash datatable
This makes the table font consistent with the rest of the UI, rather
than oddly monospaced.  The table will also play nicely now with
Dash Bootstrap themes.

I had to use custom css to ensure text in each cell of the table
didn't wrap.

In order to update the table in a callback, the entire table has
to be displayed as a child of an html.Div.  This link was useful:
https://community.plotly.com/t/pass-a-pandas-dataframe-to-a-bootstrap-table-from-a-callback/50404/6