Add option to switch between dark and light modes
There are a few ways to do this.  I used the method with a
clientside callback described here:
https://hellodash.pythonanywhere.com/adding-themes/color-modes

Some general information about clientside callbacks is here:
https://dash.plotly.com/clientside-callbacks

Using a clientside callback doesn't update the figure templates
automatically -- you have to update them specifically in a callback.
This is described in the above link.  Therefore I added a way to
specify theme to the figure generation code, using the methods from
here:
https://hellodash.pythonanywhere.com/adding-themes/figure-templates

The downside of using the clientside callback is that there is a
a delay and flicker when the colour mode is changed.  This is
mentioned here:
https://community.plotly.com/t/dash-bootstrap-theme-switcher/50798/2

Alternative methods to the clientside callback are:
- include the appropriate code in a .js file in the assets/ folder:
  https://hellodash.pythonanywhere.com/adding-themes/color-modes
  I didn't do this because it sets the theme globally and is
  explicitly not recommended when you want to switch themes.
- use the Dash Bootstrap all-in-one component ThemeSwitchAIO:
  https://hellodash.pythonanywhere.com/adding-themes/theme-switch
  I didn't do this because it seemed more complicated to update the
  figure themes, as compared to using a callback, and because I
  don't use any other all-in-one components so (a) I don't know much
  about them, and (b) using a clientside callback matches the style
  of using a regular callback so I felt it made the code more
  consistent and easier to understand.
- use the Dash Dark Theme Provider:
  https://dash.plotly.com/dash-daq/darkthemeprovider
  I didn't do this because the documentation wasn't great and I also
  had trouble getting it to work.

The navbar background colour didn't update automatically, so I had
to write another callback to update it.  This was easy because the
navbar callback is still the output of a single callback (Dash has
more robust support for multiple callbacks now, but it was still
nice to not have to dive into that).  I used the 'color' option
for navbar, because the 'navbar-dark' class specified by the 'dark'
option didn't change anything.  Documentation for navbar options
is here:
https://dash-bootstrap-components.opensource.faculty.ai/docs/components/navbar/

In the process of figuring out navbar colours, I learned that you
can find theme colours by using the Inspect function in a browser
to look at different elements of the dashboard, and then use those
variables from the theme in custom.css:
https://community.plotly.com/t/access-background-color-from-dash-bootstrap-templates-inside-dash-callback/63458/7
The arrows for the carousel need to change colours depending on the
theme, so I used this method to match them to the current theme.

To choose a theme, I used this nice theme switcher to explore light
and dark modes of different themes:
https://hellodash.pythonanywhere.com