Match colour of chips to colour theme
I couldn't manage to change the outline colour in css or in the
chip options, despite the documentation saying it's possible.
The documentation says the default colour is the Mantine theme
primary colour:
https://www.dash-mantine-components.com/components/chip
so I changed the primary colour of the theme for the chips using
the documentation here:
https://www.dash-mantine-components.com/components/mantineprovider

Unfortunately this can only be used with Mantine colours, so I had
to hardcode in which one to use.  It's a bit too light for dark
mode...

A list of Mantine colours is here:
https://v5.mantine.dev/theming/colors/