Remove border from chips
Using !important in the css gets rid of the default primary-colour
border around chips.  However I couldn't find a way to replace it
with a new border; I was only able to remove it by changing any of
its properties.  At least this way I can remove the Mantine theme
stuff when the chips are created (where I could only hardcode the
border colour, not get rid of it or use a Bootstrap theme colour).