Use custom colours for plots
This was really hard to figure out.  Updating the layout after a
figure was created didn't update all aspects of the figure, e.g.
colours.  For some reason, changing the default template before
creating the figure seems to work.

I thought that not using update_layout() might break scatterplot
selection, but this does not seem to happen, so provisional yay.