Build MongoDB database from json or odml files.
Note that this functionality shares the following files with the
user interface part of beaverdam:
- config.toml
- metadatasource.py

I added functionality to both files for database-building, because
it made sense to put all configuration parameters in a single file
(config.toml) and to put database methods into the database class
(metadatasource.py).

It might make sense to rename metadatasource.py now, becaue in the
building-database case it's actually the metadata destination.

This is the first commit where all the beaverdam functionality is
present, yay!