Use odML file name as json file name.
This removes the functionality to search for a session name in an
odML file, because session names were specific to the V4A project.
The code now carries the original metadata filename all the way
through, and it's up to researchers to name their files in a way
that makes sense for them.