Add loading indicator
This is a compromise because I can't figure out a way to further
speed up building the plots and data table.

I put the indicator on the list of applied filters, because it's
pretty intuitive and (1) putting it over the data table wasn't
visible for large datatables (the indicator went off the screen),
and (2) putting it over the figures made the figures disappear,
so you could no longer easily see what had changed in the figure.

Documentation here:
https://dash.plotly.com/dash-core-components/loading