Limit number of table rows
It is important that we limit the number of rows in the data table,
so that it doesn't slow down loading for large data sets.  Probably
no one is going to scroll through thousands of rows anyway.

The max number of datatable rows to show at once could be put into
the config file.

The Dash Bootstrap Table class doesn't have a way to limit the
number of rows shown, so I had to go back to using the stock Dash
DataTable class.  To theme the DataTable according to the Dash
Bootstrap theme currently in use, I used some custom css as
described here:
https://hellodash.pythonanywhere.com/adding-themes/datatable

But because I want beaverdam to be as independed from other websites
as possible, I took the relevant css out of AnneMarieW's solution
and added it to custom.css.  The original css from AnneMarieW is
here:
https://cdn.jsdelivr.net/gh/AnnMarieW/dash-bootstrap-templates/dbc.min.css