Get selected checklist items directly in View
This is consistent with how other UI elements are updated in View.
No other UI element passed its associated Presenter object to
builduielements_dash.py, so it was inconsistent for this to
happen for the checklist.

Now all Presenter elements are parsed within View instead of passing
them to other functions.

Let's see if this was a good design choice ;)