Be robust to missing values when filtering datatable
If a row of the datatable was missing an entry, for example because
that metadata field didn't exist in the database for that
experiment, filtering the column containing the missing entry threw
and error.

Now lists, None, and other values are treated separately during
filtering.

This method uses Pandas apply(), which is a bit slow but doesn't
seem to have too much of an effect on update time.

I tried using some numpy functions and in the process added then
removed numpy as a requirement.  Presumably numpy is faster than
pandas, but I didn't see any functions to nicely compare a value
to a list of values.  So I removed the numpy dependency, but in
the process of adding/removing I guess Poetry updated the
dependencies listed in the lock file.  Nothing meaningful should
have changed (I hope).