Make boolean filter criteria show up in chips [INCOMPLETE]
Booleans need to be converted to strings to show up in the list of
applied filters.

HOWEVER, now that they're strings, they need to be converted back
to booleans to correctly remove filter criteria by clicking on
chips.  This is complicated, because unlike filter checkboxes chips
don't differentiate between labels and values, so you can only have
strings or numbers associated with them.