Allow bool in chips; uniquely identify filter criteria in chips
Chip text has to be numeric or string, so any types other than that
didn't show up.

To be able to show bool values in chips, they therefore had to be
converted to strings, but then it was ambiguous how to interpret
chip values.  Therefore I encoded filter criteria, value, and type
into the chip value, so that the unique value and criteria could
be recovered when the chip was clicked on and used to update filter
criteria.

This was necessary for the bool values, and also solved the
undesireable method of interpreting chip values that was used
previously, where a chip value wasn't associated with a certain
criteria so if multiple criteria had identical values it would have
been impossible to distinguish them in the callback.

Still not perfect for duplicate values, since the criteria doesn't
display on the chip, but it's a start and can be built on later :)