Rename variables with two leading underscores to have only one
Python convention is to use single leading underscores for things
that should be treated as private.  Double leading underscores
have special function to invoke name mangling, which adds the
class name to the object's name.  My intention with the leading
underscores was to denote private objects, not to name mangle.

Documentation in PEP8:
https://peps.python.org/pep-0008/#descriptive-naming-styles