Make TinyDB the default database if "type" isn't specified [BREAKING]
Previously, if "type" wasn't specified in the database section of
the config file, Beaverdam would default to a MongoDB database for
backwards compatibility.  However, this isn't ideal, because the
using MongoDB relies on external dependencies.

Using TinyDB as a default doesn't require external dependencies
beyond the required packages already installed with Beaverdam.
It will likely also be the database type most users choose,
because it is simpler than installing and configuring MongoDB.
Therefore it makes more sense to use TinyDB as the default.

The first release of Beaverdam didn't include the "type" field in
the configuration file, and only used MongoDB.  Defaulting to
TinyDB, which this commit introduces, is incompatible with the
first release of Beaverdam.  To maintain the original behaviour in
legacy code, add:
type = "mongodb"
in the database section of the configuration file.