Robustly filter booleans from pie charts
Prevously, the code relied on detecting the first boolean-like
element from a list of possible boolean-like elements.  This was
not robust, because Python treats True and 1 as equivalent, so if
the actual boolean value was the second in the list (1) rather
than the first (True) the code returned the incorrect (first)
option.

The improved method checks the value and type of each possible
element, and detects elements where both value and type match the
actual boolean-like value.  This works because True has type bool,
while 1 has type int, so they can be distinguished.